Move Refill modal
