+++ src/components/ui/modals/GaslessVoting/RefillGasTankModal.tsx
@@ -4,14 +4,14 @@ import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useBalance } from 'wagmi';
 import * as Yup from 'yup';
-import { useValidationAddress } from '../../../hooks/schemas/common/useValidationAddress';
-import { useDaoInfoStore } from '../../../store/daoInfo/useDaoInfoStore';
-import { BigIntValuePair } from '../../../types';
-import { formatCoinUnits } from '../../../utils/numberFormats';
-import { BigIntInput } from '../forms/BigIntInput';
-import { CustomNonceInput } from '../forms/CustomNonceInput';
-import LabelWrapper from '../forms/LabelWrapper';
-import { AssetSelector } from '../utils/AssetSelector';
+import { useValidationAddress } from '../../../../hooks/schemas/common/useValidationAddress';
+import { useDaoInfoStore } from '../../../../store/daoInfo/useDaoInfoStore';
+import { BigIntValuePair } from '../../../../types';
+import { formatCoinUnits } from '../../../../utils/numberFormats';
+import { BigIntInput } from '../../forms/BigIntInput';
+import { CustomNonceInput } from '../../forms/CustomNonceInput';
+import LabelWrapper from '../../forms/LabelWrapper';
+import { AssetSelector } from '../../utils/AssetSelector';
 
 interface RefillGasFormValues {
   inputAmount?: BigIntValuePair;
